requirements.txt
fastapi==0.121.2
uvicorn[standard]==0.38.0
sqlalchemy==2.0.44
aiosqlite==0.21.0
pydantic==2.12.4
python-dotenv==1.2.1
pyyaml==6.0.3
requests==2.32.5
psutil==7.1.3
jinja2>=3.1.6
google-genai==1.51.0
beautifulsoup4==4.14.2
python-multipart==0.0.20
scikit-learn==1.5.2
sentence-transformers==3.1.1
transformers>=4.53.0
torch>=2.6.0
